sched/fair: Do not enable fast new idle balance if the domain is busy
If every CPU has at least 1 running task then there is no need to
enable fast new idle balance(ILB_FAST). Because if the sched domain
is busy, the normal idle load balance should try its best to pull
tasks to achieve balance. Set the total_load field to 0 which is used
as an indicator to not trigger ILB_FAST.

Signed-off-by: Chen Yu <yu.c.chen@intel.com>