sched/fair: Calculate the scan depth for idle balance based on system utilization
When CPU is about to enter idle, it invokes newidle_balance() to pull
some tasks from other runqueues. Although there is per doman
max_newidle_lb_cost to throttle the newidle_balance(), there is still
optimization room. The problem is that, there is no limitation for
newidle_balance() to launch this balance simultaneously on multiple CPUs.
Since each newidle_balance() has to traverse all the CPUs to calculate
the statistics one by one, this total time cost on newidle_balance()
could be O(n^2). This is not good for performance nor power saving.

For example, sqlite has spent quite some time on newidle balance():
6.69%    0.09%  sqlite3     [kernel.kallsyms]   [k] newidle_balance
5.39%    4.71%  sqlite3     [kernel.kallsyms]   [k] update_sd_lb_stats

Based on this observation, limit the scan depth of newidle_balance()
on LLC that,  when the LLC is under-loaded, newidle_balance() spends
less time to search for the busiest group. On the other hand, when
the LLC is overloaded, newidle_balance() tries its best to find the
busiest group.

The reason is that, when LLC is under-loaded:
1. There is less likely an emergency requirement to
   do the balance immediately. And the periodic load
   balance could help later.
2. There could be a lot of CPUs running newidle_balance()
   at the same time if the system frequently becomes idle.
   The cost on newidle_balance() would be significant with
   the increasing number of CPUs,

On the other hand, when the LLC is overloaded:
1. There is a strong requirement to eliminate the imbalance
   to offload the burden of overloaded CPUs immediately.

Based on that, let the number of scanned groups be a linear
function of the utilization ratio:

nr_groups_to_scan = nr_groups * util_ratio

Besides, save the total_load, total_capacity of the current
sched domain in each periodic load balance. This statistic
can be reused later by CPU_NEWLY_IDLE load balance if it quits
the scanning earlier. Introduce a sched feature ILB_UTIL to
control this.

Suggested-by: Tim Chen <tim.c.chen@intel.com>
Signed-off-by: Chen Yu <yu.c.chen@intel.com>