sched/fair: Reuse the group's statistic during load balance
The load balance(especially the newidle load balance) launches
the balance on multiple CPUs cocurrently, and causes O(n^2) overhead
when calculating the statistics one by one.

Peter suggested to reuse the statistic of this group if someone in the
same group has already calculated the statistic for us[1].

In theory this statistic could be also used by remote sched groups
in this domain, but that requires the current CPU to access other
CPUs' sched domain, which might bring overhead.

Link: https://lore.kernel.org/lkml/20230621111721.GA2053369@hirez.programming.kicks-ass.net/

Suggested-by: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: Chen Yu <yu.c.chen@intel.com>