sched/fair: Reuse the group's statistic during load balance
The load balance(especially the newidle load balance) launches
the balance on multiple CPUs cocurrently, and causes O(n^2) overhead
when calculating the statistics one by one.

Peter suggested to reuse the statistic of this group if someone in the
same group has already calculated the statistics[1]. For example:

       +--------------------+
       |                    |
groups |      domain 0      |
  +----+                    |
  |    +--------------------+
  |                    |child
  |                    |
 +v--------+    +------v----+         +------------------+
 |         |    |           | shared  |                  |
 | group 0 |    | domain 00 +-------->+ group statistics |
 |         |    |           |         |                  |
 +---------+    +-----------+         +------------------+
 cpu0,cpu1

                 CPU 0

In above figure, load balance is triggered on CPU0 and currently on
domain 0. Domain 0 is the parent of domain 00, and group 0 is the
local sched group under domain 0. CPU0 and CPU1 are in the same
group 0. The load balance is about to calculate the statistic of
group 0. It first checks if there is a snapshot saved in
domain 00's shared field. If yes, load it directly. Otherwise
calculate the statistic of group 0. Once the calculation is done,
the data of group 0 is stored in domain 00's shared field. A short
time later if the load balance happens on CPU1 and also the same
level of domain 0, it can load this snapshot without having to
re-calculate it.

There is a restriction that, only 1 CPU can update the snapshot
at the same time. And when the CPU is updating the snapshot,
other CPUs are not allowed to read from the snapshot. Multiple
readers are allowed to read the snapshot at the same time. In
theory the RCU is a good candidate. But RCU mechanism needs to
leverage call_rcu() to be noticed of the end of reading, and
update the snapshot then. The RCU version will be tried later.

Link: https://lore.kernel.org/lkml/20230621111721.GA2053369@hirez.programming.kicks-ass.net/ #1
Suggested-by: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: Chen Yu <yu.c.chen@intel.com>