sched/fair: Do not trigger the idle load balance if a timer is not far away
If there is a timer to be triggered shortly, there is no need to launch
the idle load balance. This is because that, a timer event usually
accompany with a task wakeup. In order to benefit the wakeup latency do
not launch the time costly idle load balance on this CPU.

To be more specific, compare the expire time of the latest timer and the
cost to finish a idle load balance, if the former is shorter, do not
launch the idle load balance.

To find the latest timer, two places should be checked:
1. The expired time alreayd set into the clock event.
2. The timers queued and the leftmost one in that tree, whose
   expired time is about to be set into the clock event.

Since logic 2 could be time costly, only consider logic 1 to do a
fast lookup.

Suggested-by: Tim Chen <tim.c.chen@intel.com>
Suggested-by: Arjan van de Ven <arjan.van.de.ven@intel.com>
Suggested-by: Len Brown <len.brown@intel.com>
Signed-off-by: Chen Yu <yu.c.chen@intel.com>