sched/fair: Introduce the LB_SNAPSHOT to reuse the group statistic
The load balance(especially the newidle load balance) launches
the balance on multiple CPUs cocurrently, and causes O(n^2) overhead
when calculating the statistics one by one.

Peter suggested to reuse the statistic of this group if someone in the
same group has already calculated the statistics[1]. For example:

       +--------------------+
       |                    |
groups |      domain 0      |
  +----+                    |
  |    +-------------------+
  |     ^              |child
  |     |sd            |
 +v--------+    +------v----+         +------------------+
 |         |    |           | shared  |                  |
 | group 0 |    | domain 00 +-------->+ group statistics |
 |         |    |           |         |                  |
 +---------+    +-----------+         +------------------+
 cpu0,cpu1

                 CPU 0

In above figure, load balance is triggered on CPU0 and it is on
domain 0. Domain 0 is the parent of domain 00, and group 0 is the
local sched group under domain 0. CPU0 and CPU1 are in the same
group 0. The load balance is about to calculate the statistic of
group 0. It first checks if there is a snapshot saved in
domain 00's shared field. If yes, load it directly. Otherwise
calculate the statistic of group 0. Once the calculation is done,
the data of group 0 is stored in domain 00's shared field for
other CPU's use.

Introduce sched feature LB_SNAPSHOT to control it.

For the data structure, according to the suggestion from Qiuxu and
Tim, there is a per sched_domain_shared ring buffer with a default size
of 4. The writer saves the statistic in the snapshot, and then
tells the reader to load from the latest snapshot. Something like:

        reader  ---> wrap move
         +
         |
         v
+--------+---------+---------+---------+
|        |         |         |         |
|        |         |         |         |
+------------------+---------+---------+
                   ^
                   |
                   +
                  writer   ------> wrap move

Limitations:

The writer who saves the snapshot could overwrite the data that
the readers are accessing. This happens when the writer wraps in
the ring buffer and catches up with the reader. At that time the
reader is still reading the old data, and has not switched to the
latest data yet. This is caused by that there is no strict
synchronization between the writer and the reader, due to effeciency
consideration. The bad scenario is descibed as:

       reader  ---> wrap move
         +
         |
         v
+------------------+---------+---------+
|        |         |         |         |
|        |         |         |         |
+------------------+---------+---------+
         ^
         |
         +
        writer   ------> wrap move

However, above race window could be small. If it really happens, it
should not be a big deal. Because update_sd_lb_stats() itself calculates
the statistic without any rq lock held that, the statistic calculation
is not 'atomic'.

Besides, with LB_SNAPSHOT enabled, there are extra memcpy operations
during the statistic load/save. If there is no frequent newidle balance
triggered, this extra memcpy operation might not offset the benefit
of reusing the snapshot. But consider that in the scenario the newidle
balance is not the bottleneck, this extra cost could be neglected.

Link: https://lore.kernel.org/lkml/20230621111721.GA2053369@hirez.programming.kicks-ass.net/ #1
Suggested-by: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: Chen Yu <yu.c.chen@intel.com>