sched/fair: Do not launch newidle balance if the domain is balanced
It is observed that the newidle balance is launched many times but
fails to pull any task to the idle CPU. On a system with large number
of CPUs this wastes a lot of time and gain nothing. Run the netperf
on a platform with 112 CPUs/1 Socket online:

for i in $(seq 1 10); do
	netperf -4 -H 127.0.0.1 -t TCP_RR -c -C -l 100 &
done

Total statistics on all CPUS on domain1(LLC domain), in 5 seconds:
.domain1.CPU_NEWLY_IDLE.lb_count': 1085672
.domain1.CPU_NEWLY_IDLE.lb_balanced: 831980
.domain1.CPU_NEWLY_IDLE.lb_failed: 251269
.domain1.CPU_NEWLY_IDLE.lb_imbalance: 253693
.domain1.CPU_NEWLY_IDLE.lb_nobusyg: 367297
.domain1.CPU_NEWLY_IDLE.lb_nobusyq: 464682
.domain1.CPU_NEWLY_IDLE.lb_sg_scan: 59711907

Among 1085672 newidle balance attempt, 831980 is already balanced,
the fail rate is up to 76.63%. The total sched groups scanned is
59711907, and 76.63% of these scan is a waste of time. This is
not good for performance nor power.

If the newidle balance fails to pull any tasks, do not let this
CPU to launch newidle balance on this domain within 1 msec.

Signed-off-by: Chen Yu <yu.c.chen@intel.com>