sched/fair: Do not launch the newidle balance if the sleeping task will be woken up soon
If a task is the only running task of the CPU, and if that task falls asleep,
the CPU becomes idle. This idle CPU tries to pull some tasks from other
busy runqueues. The time spent on pulling tasks could be costly, because
newidle load balance has to traverse all the CPUs in the domain to
find the busiest one. If the task will be woken up soon, there is
no need to waste the time to find another task, just let the CPU stay
idle, and when the task is woken up again, it could reuse its previous
CPU to improve cache locality.

Suggested-by: Tim Chen <tim.c.chen@intel.com>
Signed-off-by: Chen Yu <yu.c.chen@intel.com>