sched/fair: Record the average sleep time of a task
There is a requirement to have the information on how
long a task sleeps. This information can be used for
better task placement that, the short sleeping task
can be woken up on its previous idle CPU easier. The
optimization will be introduced in the next patch.

Define the sleep time of the task as:
The time delta between this task was dequeued and enqueued.

The update_avg() is approximately averaging over recent 8
sample data. So the average sleep time of this task reflects
the latest 8 sleep duration.

Signed-off-by: Chen Yu <yu.c.chen@intel.com>