sched/fair: Choose cache hot idle CPU if no cache cold one is found
The motivation of SIS_CACHE:
A cache hot idle CPU is defined as such an idle CPU that, once
there was a task running on it not long ago. This kind of idle
CPU usually has the hot data in the cache for that previous
running task. So if that task is going to be woken up soon,
it is better to put that task on its previous running CPU.

During task wakeup, SIS_CACHE skips the idle CPUs that are
cache hot for other tasks, and waits for the task once running
on it. This could make the wakee easier to choose its previous
CPU.

However there is a problem that, if all the idle CPUs are cache
hot, the select_idle_cpu() would choose the target CPU, which
might be less optimized. Instead, the first cache hot CPU should
be chosen, because it is an idle CPU afterall.

With current optimization applied, the benchmark score w/ vs w/o
SIS_CACHE is as followed. There are obvious improvements for
netperf/tbench under 75% and 100% load. Others do not have any
significant impact consider the run-to-run variance.

Suggested-by: K Prateek Nayak <kprateek.nayak@amd.com>
Signed-off-by: Chen Yu <yu.c.chen@intel.com>