sched/fair: Record the short sleeping time of a task
During task wakeup, the wakee firstly checks if its previous
running CPU is idle. If yes, choose that CPU as its first
choice. However, in most cases, the wakee's previous CPU
could be chosen by someone else, which breaks the cache
locality.

If a task is short-sleeping one, there is no need to
migrate it to other CPUs. Record the average short
sleeping time of the task for this purpose.

The sleeping time is defined as the delta between this
task being dequeued and enqueued. Only the sleep time
shorter than sysctl_sched_migration_cost will be recorded.
If the sleep time is longer than sysctl_sched_migration_cost,
give a penalty to the task by shrinking the sleeping time to
half.

Later this sleeping time can be used for better task placement
during task wake up.

Suggested-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Signed-off-by: Chen Yu <yu.c.chen@intel.com>