sched/fair: Calculate the cache-hot time of the idle CPU
When a CPU is about to become idle due to task dequeue, calculate
the possible average idle time of this CPU, and treat this period
as cache hot. This information can be used to preserve the idle CPU
for a short while, so when that task is woken up again, it can choose
its previous CPU for better cache locality.

This is the preparation for the next patch to introduce SIS_CACHE
based task wakeup.

Signed-off-by: Chen Yu <yu.c.chen@intel.com>