sched/fair: skip the cache hot CPU in select_idle_cpu()
Problem statement:
When task p is woken up, the scheduler leverages select_idle_sibling()
to find an idle CPU for it. p's previous CPU is usually a preference
because it can improve cache locality. However in many cases, the
previous CPU has already been taken by other wakees, thus p has to
find another idle CPU.

Proposal:
Inspired by Mathieu's idea[1], introduce the SIS_CACHE. It considers
the sleep time of the task for better task placement. Based on the
task's short sleeping history, keep p's previous CPU idle for a short
while. Later when p is woken up, it can choose its previous CPU in
select_idle_sibling(). When other task is woken up, it checks if the
idle CPU it finds is cache-hot or not. If it is a cache-hot one,
tries to find the next idle CPU. Repeat this until half of the suggested
number of CPUs to scanned is reached. This increases the ratio that the
previous running CPU is picked by the wakee.

This does not break the work conservation of the scheduler, because
wakee will still try its best to find an idle CPU. The difference is that
different idle CPUs might have different priorities.

Signed-off-by: Chen Yu <yu.c.chen@intel.com>