sched/fair: Calculate the cache-hot time of the idle CPU
When a CPU is about to become idle due to task dequeue, uses
the dequeued task's average cache hot duration to set the cache
hot timeout of this idle CPU. This information can facilitate
SIS to skip the idle cache-hot CPU and scan for the next one.
When that task is woken up again, it can choose its previous
CPU and gets better cache locality.

This is a preparation for the next patch to introduce SIS_CACHE
based task wakeup.

Signed-off-by: Chen Yu <yu.c.chen@intel.com>