sched/fair/HACK: Add sysctl to control the number of cache hot idle CPUs
echo xx > /proc/sys/kernel/nr_div

xx is a positive value. The max number of cache hot idle CPUs is
nr/xx. For example, if xx = 2, and the nr is the number suggested
by SIS_UTIL, then the max number of cache hot idle CPUs is nr / 2.
This is a hack to restrict the scan depth.

Signed-off-by: Chen Yu <yu.c.chen@intel.com>