fs/resctrl: Cleanup as L3 is no longer the only monitor resource
The feature to sum event data across multiple domains supports systems
with Sub-NUMA Cluster (SNC) mode enabled. The top-level monitoring files
in each "mon_L3_XX" directory provide the sum of data across all SNC
nodes sharing an L3 cache instance while the "mon_sub_L3_YY" sub-directories
provide the event data of the individual nodes.

SNC is only associated with the L3 resource and domains and as a result
the flow handling the sum of event data implicitly assumes it is
working with the L3 resource and domains.

Reading of telemetry events do not require to sum event data so this
feature can remain dedicated to SNC and keep the implicit assumption
of working with the L3 resource and domains.

Add a WARN to where the implicit assumption of working with the L3 resource
is made and add comments on how the structure controlling the event sum
feature is used.

Suggested-by: Reinette Chatre <reinette.chatre@intel.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>