x86/resctrl: Add support for L3 occupancy monitoring via RMID MMIO read
The CMRC ACPI table (Cache Monitoring Registers for CPU Agents Description)
contains the MMIO address for reading the LLC occupancy. If ERDT is enabled
on this platform, replace the MSR read operation with a read from the MMIO
address.

Introduce erdt_mon_read(), a helper to retrieve monitoring data for a
given RMID and event ID from an ERDT domain. Initial support is added
for the L3 occupancy monitoring event (QOS_L3_OCCUP_EVENT_ID).

The function locates the MMIO address for the target RMID using clump
and stride information from the CMRC structure and reads the occupancy
value. The result is scaled using the up_scale factor defined in the
firmware table.

The logic for computing the MMIO address and reading the value is
encapsulated in a new internal helper function, erdt_read_l3_occupancy().

This provides the foundation for future monitoring features when region
aware RDT is enabled.

Co-developed-by: Anil S Keshavamurthy <anil.s.keshavamurthy@intel.com>
Signed-off-by: Anil S Keshavamurthy <anil.s.keshavamurthy@intel.com>
Signed-off-by: Chen Yu <yu.c.chen@intel.com>