x86/resctrl: Extract a function for MBA to write to specific type of region aware RDT
Currently in the erdt_ctrl_update(), the MBA Optimal BW Register
is only touched. However, there are MBA Minimal and Max BW Registers
that should be adjusted accordingly. Introduce a new helper function
to write to the MBA register according to the input parameter
of the MBA types.

No functional change expected.

Signed-off-by: Chen Yu <yu.c.chen@intel.com>