x86/resctrl: Fix the region aware MBA settings for Optimal Cap
According to the RDT spec:
"The Optimal Cap should be programmed to be between or equal to
the Min andMax Cap levels. The Maximum Cap should be programmed
to be greater than or equal to the Minimum and Optimal caps.
The default hardware state is initialized with Max BW Cap == Optimal BW
==Min BW Cap (and throttling is disabled by default).
"

Since the method for exposing the interfaces to user space is still
under discussion, temporarily set Min to 1 and Max to 255 to comply
with the spec.

Reported-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Chen Yu <yu.c.chen@intel.com>